Add files via upload
Used car market is huge and growing more than ever in India. Anyone who has bought or sold a used car will agree that one of the biggest challenge in the trade is estimating a fair and market smart price which would benefit both buyer and the seller. Consequently, other than individuals, automobile manufactures, used cars dealers and loan lending banks would also be interested in a model which would help them predict right price of the car.

All the parties mentioned above may have a different motivation or expectation from the model. A seller is trying to maximize the cost while the buyer is trying to minimize. Car manufactures are interested in estimating the competition they face from the used car market while the used car dealers are trying to maximize their cut from the deal. Finally banks are estimating the right loan value and the associated risk on the loan compared to the returns.

Other than regular buy and sell, car lease business model is also very prevalent and helps individual to obtain a car by paying monthly “rent” on the car instead of buying it.

Considering the complex dynamics of used car sales business, a machine learning model could be appropriate to estimate the price of a car.

You are expected to do independent search in understanding the factors and the business in little more detail and predict the price of used vehicles based on historical data provided. Also study the data which is provided to you and make your own notes on how you can use the data most effectively to predict the price of the car.